+++ easycv/models/backbones/conv_mae_vit.py
@@ -8,8 +8,9 @@ import torch.nn as nn
 from timm.models.layers import trunc_normal_
 
 from easycv.models.registry import BACKBONES
+from easycv.models.utils import DropPath
 from easycv.models.utils.pos_embed import get_2d_sincos_pos_embed
-from .dynamic_vit_transfomer import Block, DropPath
+from .dynamic_vit_transfomer import Block
 
 
 class PatchEmbed(nn.Module):
